+++ platform/ios/src/MGLMapView.h
@@ -355,9 +355,9 @@ MGL_EXPORT IB_DESIGNABLE
 /**
  A Boolean value indicating whether the map should prefetch tiles.
  
- Loads tiles at a lower zoom-level to pre-render a low resolution preview while more detailed tiles are loaded.
+ When this property is set to YES, the map view prefetches loads tiles designed for a low zoom level and displays them until receiving more detailed tiles for the current zoom level. The prefetched tiles typically contain simplified versions of each shape, improving the map view’s perceived performance.
  
- Default is YES.
+ The default value of this property is YES.
  */
 @property (nonatomic, assign) BOOL prefetchesTiles;