+++ platform/ios/src/MGLMapView.mm
@@ -2631,15 +2631,11 @@ public:
 }
 
 - (void)setPrefetchesTiles:(BOOL)prefetchesTiles{
-    if(!prefetchesTiles)
-    {
-        _mbglMap->setPrefetchZoomDelta(0);
-    }
-    else
-    {
-        //Reset to default value.
-        _mbglMap->setPrefetchZoomDelta(4);
-    }
+    _mbglMap->setPrefetchZoomDelta(prefetchesTiles ? mbgl::util::DEFAULT_PREFETCH_ZOOM_DELTA : 0);
+}
+
+- (BOOL)prefetchesTiles{
+    return _mbglMap->getPrefetchZoomDelta() > 0 ? YES : NO;
 }
 
 #pragma mark - Accessibility -